Tabletop: Handle rotation
Convert everything over to consider external co-ordinates in the range -1.0...1.0 rather than pixel values, thus all rotation to portrait is doing is uncovering part of the play space that was previously outside the landscape bounds.

Just toggle hidden on a statsView, and use layoutSubviews to position them; with setNeedsLayout() being the "hey, I just moved a thing" as it should be.