Core Data: clean up constructors
Where there's a non-optional to-one relationship with a parent (Monster, Lair, Combatant), have that passed in with the constructor (which does the right thing for ordered sets without faffing through a setter).

Exception is Lair which I intend to make the monster relationship to-many at some point, so don't change the constructor right now.