Monsters View: use NSFetchedResultsController's knowledge of section names
The crash is happening because the section might not have fetched the monsters yet; but since the controller has cached the section name fields anyway, pull those out. For the section title we translate that into something with CR & XP; for indexes we override completely for CR so we get our cute 1/4 symbols.

Fixes #2, and fixes #3