Monster: parse Languages field
Yay, the last string field (at least aside from actions/attacks) is now parsed.

Languages it turns out has a very simple coverage for 80% of the monsters (list of spoken or understood languages) and then a complex Option case for the minority. Opted to go for a table of languages (like tags) rather than an enum.

Closes #70