Markup: make parseText public
Now that it accepts a full attributes set, and features list, this is perfect for a public function with just one modification - make sure it never refers to mutableText and always just returns the attributed string.

This is fortunately possible because the only reason we use mutableText is for tables, and those are block elements so above this function call.