Markup: allow inline markup in tables and headings
With the changes to parseText() this is pretty easy, and only really requires that the table columns be `NSAttributedString` rather than `String`, which surprisingly makes calculating the size easier.

It also means that we can get rid of the hacky replaceSingleQuotes() function and just make "'" an operator in parseText() that gets replaced.