Core Data: add model for Adventure and AdventureImage
No specific reason to split off image like this, but the class encapsulates some clever tricky with storing the actual image data on the filesystem and managing it alongside the core data object. Thus it makes sense to break it out like this, then I can always use it for other things later.