Extended Navigation Bar: handle scroll view insets
Rather than anchoring a table view to the bottom of the extended navigation bar, instead allow it to extend underneath as normal, and then adjust the insets and offsets to compensate as the extension is shown and hidden.

Also internalize removing and restoration of the navigation bar shadow as we move to and from windows.

Fixes #6