Game: add Players
A game's players are inherited from the adventure, but can be independently adjusted (since a player might not show up, or a player might come who hasn't in the past, or a new player might need to be created, etc.) Players created in the game are added to the adventure on the default assumption that they might make it to the next one.

For now the player cells in the game show the player name and the PP, because this is the screen that will tend to be used while actually running the game, and PP becomes important while things like XP don't so much.