AdventureGamesViewController: support reordering games
Not strictly that important a feature, but it gave me a good chance to figure out reordering tables so that code is lying around if I need it later. Plus there's probably a use for it if a session already has encounters in it, and we don't run it, and I want to create a new game before it, etc.