Encounter: basic UI for encounters in adventures and games
Nothing hugely fancy here, mostly just copy and paste from the existing Games and Players code.
Adventure shows Encounters that are not attached to Games. There's a quick button for creating a new encounter because that's a common operation and I don't want to have to go into Editing mode to do that (it's okay to do that to delete them, that's uncommon).
Game shows Encounters attached to that Game, and in edit mode shows the set from the Adventure too with (+) buttons. Since a new Game starts in Edit mode, I figure this is the right time to add in Encounters. There's also a shortcut to add a new encounter directly to the game for on-the-fly encounters. Not sure whether the "add encounter from adventure after creating the game" use will be common or not. time will tell.