Implement app to show Initiative Order
Changes to Dungeon Master itself are adding a new NetworkController class that handles broadcasting the changes to the "current encounter" (which I realized I have no concept of in the model!)

Initiative Order app is pretty simple and just receives these and renders appropriately.

DungeonKit gains a whole bunch of code for handling the "DungeonNet", a network of peers, connections, and message formatting for passing things between devices. This will get extended a lot once I have player apps and stuff!