InitiativeViewController: set the round to 1 earlier
We want the network to know about the encounter at the point of initiative roll, this means setting the round to 1 at the point we do that.

But that then means we have to track we did that, so when Done is tapped, the currentTurn is set (we can't use round == 0 for that anymore).