PlayedGamesViewController: rebuild sectionsInResults map in a better way
First move it to just the cached/invalidatable property pattern that I've been using elsewhere.

Secondly invalidate in the didChangeSection delegate method, rather than at the end; that way "delete row" is called with it still cached at the old value (so the index paths are removed correctly), and "new row" is called with the cache pre-updated (so the index paths are inserted correctly).

This works because we get an didChangeRow to insert or delete in a new section for which we get didChangeSection, and we get these in the expected order.

And that means we don't actually have to do anything in editing mode, because the section is already there.

Lastly in indexPath(forIndexPathInResults:) don't do a playedGames lookup, and instead rely on the rebuilt sectionsInResults map. This ensures we then get the right result.