Skip dead monsters during turn order.
Move the "next turn" logic out of the view controller and into the model where it belongs. Skip dead monsters in the turn order, and be more careful about updating the round to deal with cases where the first and last monster are dead.

Move "is alive" into Combatant, and pass it to the Initiative Order app so that it too can make the same judgement about who has the next turn.